% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementary_functions.R
\name{fisher_exact_test}
\alias{fisher_exact_test}
\title{fisher_exact_test
one side? two side?
see https://www.pathwaycommons.org/guide/primers/statistics/fishers_exact_test/
and paper (Enrichment or depletion of a GO category within a class of genes: which test?) for more details 
This function allows you to compute fish exact pvalue of gene list for selected  pathways
         genes_in_common      genes_in_reference
pathway          a                   c
no_pathway       b                   d}
\usage{
fisher_exact_test(selected_pathways, gene_input, gene_pathway_matrix)
}
\arguments{
\item{gene_input}{A vecor of genes to be annotated. It should have same ID types(Ensembl ID, HUGO gene symbol) as the genes in \emph{gene_pathway_matrix}.}

\item{gene_pathway_matrix}{A binary background matrix whose columns are the pathways/gene sets and 
whose rows are all the genes from pathways/gene sets . It could be in sparse matrix format ((inherit from class "sparseMatrix" as in package Matrix) to save memory.
For gene i and pathway j, the value of matrix(i,j) is 1 is gene i belonging to pathway j otherwise 0.}
}
\description{
fisher_exact_test
one side? two side?
see https://www.pathwaycommons.org/guide/primers/statistics/fishers_exact_test/
and paper (Enrichment or depletion of a GO category within a class of genes: which test?) for more details 
This function allows you to compute fish exact pvalue of gene list for selected  pathways
         genes_in_common      genes_in_reference
pathway          a                   c
no_pathway       b                   d
}
\examples{
a=regression_selected_pathways(gene_input =c("TRPC4AP","CDC37","TNIP1","IKBKB","NKIRAS2","NFKBIA","TIMM50","RELB","TNFAIP3","NFKBIB","HSPA1A","NFKBIE","SPAG9","NFKB2","ERLIN1","REL","TNIP2","TUBB6","MAP3K8"),gene_pathway_matrix=gene_pathway_matrix,lambda=0.007956622,alpha=0.5)
}
